requirements.txt
# Core Dependencies
anthropic>=0.25.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.0
asyncio>=3.4.3

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.1.0

# Utilities
python-multipart>=0.0.9
jinja2>=3.1.3

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Code Quality
mypy>=1.9.0
pylint>=3.1.0
black>=24.3.0
types-pyyaml>=6.0.12

# Logging & Monitoring
structlog>=24.1.0
colorama>=0.4.6
